break out to sep components. download rights btn
